+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # adrez-mapper Revisions
+2022-03-12
+- add: url lookup translator from Facebook to code id
+
 20-09-29
 - add: code.typeGuid let the informatie be stored